# src/Esquio.UI.Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/toggles/parameters/{toggleType}": {
      "get": {
        "tags": [
          "Toggles"
        ],
        "operationId": "Toggles_Reveal",
        "parameters": [
          {
            "type": "string",
            "name": "toggleType",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/RevealToggleResponse"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/toggles/types": {
      "get": {
        "tags": [
          "Toggles"
        ],
        "operationId": "Toggles_KnownTypes",
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/KnownTypesToggleResponse"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/toggles/parameters": {
      "post": {
        "tags": [
          "Toggles"
        ],
        "operationId": "Toggles_AddParameter",
        "parameters": [
          {
            "name": "parameterToggleRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddParameterToggleRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/toggles": {
      "post": {
        "tags": [
          "Toggles"
        ],
        "operationId": "Toggles_Add",
        "parameters": [
          {
            "name": "postToggleRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddToggleRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Users/my": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_My",
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/MyResponse"
            }
          },
          "403": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_List",
        "parameters": [
          {
            "type": "integer",
            "name": "pageIndex",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "pageCount",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListUsersResponse"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Add",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddPermissionRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Update",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePermissionRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Users/{subjectId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Details",
        "parameters": [
          {
            "type": "string",
            "name": "subjectId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DetailsUsersResponse"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Users/{subjectid}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Delete",
        "parameters": [
          {
            "type": "string",
            "name": "subjectId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": {}
        }
      }
    },
...
    "ValidationProblemDetails": {
      "allOf": [
        {
          "$ref": "#/definitions/ProblemDetails"
        },
        {
          "type": "object",
          "properties": {
            "errors": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      ]
    },
...
    "RevealToggleResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "parameters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RevealToggleParameterResponse"
          }
        }
      }
    },
    "RevealToggleParameterResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "clrType": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "KnownTypesToggleResponse": {
      "type": "object",
      "required": [
        "scannedAssemblies"
      ],
      "properties": {
        "scannedAssemblies": {
          "type": "integer",
          "format": "int32"
        },
        "toggles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/KnownTypesToggleDetailResponse"
          }
        }
      }
    },
    "KnownTypesToggleDetailResponse": {
      "type": "object",
      "properties": {
        "assembly": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "friendlyName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "AddParameterToggleRequest": {
      "type": "object",
      "properties": {
        "productName": {
          "type": "string"
        },
        "featureName": {
          "type": "string"
        },
        "toggleType": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "AddToggleRequest": {
      "type": "object",
      "properties": {
        "productName": {
          "type": "string"
        },
        "featureName": {
          "type": "string"
        },
        "toggleType": {
          "type": "string"
        },
        "parameters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddToggleRequestDetailParameter"
          }
        }
      }
    },
    "AddToggleRequestDetailParameter": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "MyResponse": {
      "type": "object",
      "required": [
        "isAuthorized",
        "readPermission",
        "writePermission",
        "managementPermission"
      ],
      "properties": {
        "isAuthorized": {
          "type": "boolean"
        },
        "readPermission": {
          "type": "boolean"
        },
        "writePermission": {
          "type": "boolean"
        },
        "managementPermission": {
          "type": "boolean"
        }
      }
    },
    "ListUsersResponse": {
      "type": "object",
      "required": [
        "total",
        "pageIndex",
        "count"
      ],
      "properties": {
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "pageIndex": {
          "type": "integer",
          "format": "int32"
        },
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListUsersResponseDetail"
          }
        }
      }
    },
    "ListUsersResponseDetail": {
      "type": "object",
      "required": [
        "managementPermission",
        "readPermission",
        "writePermission"
      ],
      "properties": {
        "subjectId": {
          "type": "string"
        },
        "managementPermission": {
          "type": "boolean"
        },
        "readPermission": {
          "type": "boolean"
        },
        "writePermission": {
          "type": "boolean"
        }
      }
    },
    "DetailsUsersResponse": {
      "type": "object",
      "required": [
        "managementPermission",
        "readPermission",
        "writePermission"
      ],
      "properties": {
        "subjectId": {
          "type": "string"
        },
        "managementPermission": {
          "type": "boolean"
        },
        "readPermission": {
          "type": "boolean"
        },
        "writePermission": {
          "type": "boolean"
        }
      }
    },
    "AddPermissionRequest": {
      "type": "object",
      "required": [
        "read",
        "write",
        "manage"
      ],
      "properties": {
        "subjectId": {
          "type": "string"
        },
        "read": {
          "type": "boolean"
        },
        "write": {
          "type": "boolean"
        },
        "manage": {
          "type": "boolean"
        }
      }
    },
    "UpdatePermissionRequest": {
      "type": "object",
      "required": [
        "read",
        "write",
        "manage"
      ],
      "properties": {
        "subjectId": {
          "type": "string"
        },
        "read": {
          "type": "boolean"
        },
        "write": {
          "type": "boolean"
        },
        "manage": {
          "type": "boolean"
        }
      }
    }
  }
}